Crate gomspace_p31u_service

Kubos Service for interacting with [GomSpace p31u EPS]

Configuration

The service must be configured in /home/system/etc/config.toml with the following fields:

[gomspace-p31u-service.addr]

ip - Specifies the service's IP address

port - Specifies the port on which the service will be listening for UDP packets

[gomspace-p31u-service]

bus - Specifies the I2C bus the antenna system is connected to

i2c_addr - Specifies the I2C address of the antenna system's primary microcontroller

Example: [gomspace-eps-service.addr] ip = "0.0.0.0" port = 8021

[gomspace-eps-service]
bus = "/dev/i2c-0"
i2c_addr = "0x08"

Starting the Service

The service should be started automatically by its init script, but may also be started manually

Structs

GenericResponse
Common response fields structure for requests which don't return any specific data

RawCommandResponse
Response fields for 'issueRawCommand' mutation

Enums

AckCommand
Return field for 'ack' query